+++ src/js/components/pagination.js
@@ -92,7 +92,7 @@
             this.element.empty();
 
             // Generate Prev link
-            if (o.lblPrev) this._append(o.currentPage - 1, {text: o.lblPrev});
+            if (o.lblPrev) this._append(this.currentPage - 1, {text: o.lblPrev});
 
             // Generate start edges
             if (interval.start > 0 && o.edges > 0) {
@@ -126,7 +126,7 @@
             }
 
             // Generate Next link (unless option is set for at front)
-            if (o.lblNext) this._append(o.currentPage + 1, {text: o.lblNext});
+            if (o.lblNext) this._append(this.currentPage + 1, {text: o.lblNext});
         },
 
         _append: function(pageIndex, opts) {